requirements/requirements.txt
############################################
# Dependencies
############################################
schedule>=1.1.0,<1.2
python-dotenv>=0.21.0,<0.30
lxml>=4.9.1,<5.0
beautifulsoup4>=4.11.1,<4.20
requests>=2.28.1,<2.30
git+https://github.com/leonardcser/selenium-scheduler
